+++ s3cortxutils/s3cipher/s3cipher/s3cipher
@@ -2,7 +2,7 @@
 
 import sys
 import os
-from s3backgrounddelete.cortx_s3_cipher import CortxS3Cipher
+from s3cipher.cortx_s3_cipher import CortxS3Cipher
 
 # S3 wrapper for LDR security
 s3cipher = CortxS3Cipher()
